sensors_data/requirements.txt
paho-mqtt>=1.6.1
pandas>=1.5
